Randomly select "new villager" or "upgrade room" quest
TODO: Make upgrade quests easier to understand:

Upgrade quests only work if the "from" room was built to satisfy a quest.
For example, if the only active quest you have is "crafting room -> workshop",
and you build a new crafting room and then upgrade it to a workshop, that
won't complete the quest. If you had TWO active quests, one for a crafting
room and one for a workshop, and then you built a crafting room (thus
satisfying a quest) and then converted it to a workshop, THAT would satisfy
the workshop quest.

Also, if you have a quest to upgrade a crafting room to a deluxe workshop,
upgrading a (basic) workshop to a deluxe workshop will not complete the
quest. This might not be very intuitive.

Also, if you have an active "crafting room -> deluxe workshop" quest, and you
simply build a brand-new deluxe workshop from scratch, that WILL satisfy the
quest.  This is either a bug, or is the way things should work. i.e. maybe
upgrade quests simply give you the OPPORTUNITY to "phase out" and old room in
favor of an upgrade, but you can also just leave that room alone and either
build the upgraded room, or upgrade a different kind of room (or a non-quest
room) to satisfy the quest.

Lots to think about here. But, technically, things are tolerable with the
existing logic.