Baker will prefer to fill their inventory for work
If there are multiple available ovens, the baker should completely
fill their inventory with ingredients before heading to the jobsite.

Prior to this commit, bakers would gather enough supplies for ONE
bread oven, and then head to the jobsite. This resulted in a lot of
back and forth to the storeroom.

TODO:
- Bread ovens set their "baking start time" to the current "hour".
  However, hours are quite long. If the baker adds bread at "8:59",
  the bread will be ready at "9:00".  We should improve that.