Add config option to ENABLE around-flag door scanning
During play-testing, I stumbled across a bug with room detection
that seemed to cause an infinite loop. The server crashed due to
excessive tick duration.

Since I know the risk is out there, and because I've been leaning
toward "registered" doors as a gameplay pattern, I'm disabling
the automatic room detection by default. It can be toggled on if
you want to chance it.