Updated quest generation to happen in advance
Rather than scrambling to generate quests first thing in the
morning, right after the new villager has been added to the
village - we simply spend 50 otherwise "free" ticks generating
the batch of quests so they are ready when a new villager is
added to town. That batch gets claimed by the new village and
we then start using more "free" ticks to precompute for the
next person.

Also:
- Some better const naming throughout
- Changed the particle effect when a room is changed
- Defined a block cost for cauldron

TODO:
- Test if gatherer and explorer still work
- Test if gatherer maps are used correctly
- Make job board last after a reboot
- Maybe: Reopen work dialog after adding work
- Figure out why farmers won't start working
- Make job seekers prioritize dropping loot