Cont: Preparing or declarative gatherer (WIP)
One challenge is going to be showing every possible loot item that
can be gathered in the job board.

Can probably use this project as a reference:
- https://github.com/way2muchnoise/JustEnoughResources/blob/1.18.2/src/main/java/jeresources/entry/DungeonEntry.java

Failing that, we can always just introduce a "loot from map" meta-
item that can be selected on the job board, instead of the specific
item. But let's give this a go anyway.

TODO:
- Test if gatherer maps are used correctly
- Make job board last after a reboot
- Maybe: Reopen work dialog after adding work
- Figure out why farmers won't start working
- Make job seekers prioritize dropping loot