Huge, untested refactor - time warp for all jobs
General philosophy: every job has its own "warper" and we
grab that warper for each villager during the time warp
phase of the town flag's loading process.

Each warper simulates the passage of time and then returns
an updated town state which is passed into the next warper
as the starting point.