Better handling of multi-villager warps
First, we calculate the specific ticks where the work for each villager
would be run and store a tuple of [tick, lambda function] in a list.
Finally, we sort that list by tick and run each lambda in the sorted
list. This should mean that each villager action (lambda) is run in
the order that it would be run if this were happening in real-time.

Some light play testing shows that this seems to be working well.