Forward port to 1.19
Add icons for every villager tab

+ fix click targets

Fix bug causing new worlds to crash on room creation

Introduce new code layout to ensure initialization

Whenever new NBT data is added, we tend to have an "initializer" for the
corresponding data store in the town flag.

This new layout ensures that for every NBT deserializer, there is a
correponding initializer that can be run when the flag is first placed.