WIP: Found the bug. Fixing it.
Here's how it plays out:
- Villager inserts first item
- Wants items for next stage
- No items for next stage
- Only 100 ticks need to pass
- Gives up on working

This is too fast. We set it this low to allow workers to cycle through
all their job options after going to the job board.

But if they inserted at least one item, we should wait a lot longer to
give the player a chance to understand what items are required.

FIXME:
- Jobs are definitely broken by recent refactors
  - Worker does not transition from inserting first ingredient to
  collecting tools for second step. They just stand there frozen.

TODO:
- Add soup cook
  - [ ] Finish and test implementation
-[ ] Write good integration tests for soup cook
   (following pattern from blacksmith)