Pretty decent production cycle (and even some testing) for soup cook
Here's how it plays out:
- Villager inserts first item
- Wants items for next stage
- No items for next stage
- Only 100 ticks need to pass
- Gives up on working

This was too fast. We set it this low to allow workers to cycle through
all their job options after going to the job board.

But if they inserted at least one item, we now wait a lot longer to
give the player a chance to understand what items are required.

TODO:
- Add soup cook
  - [x] Finish and test implementation
-[ ] Write good integration tests for soup cook
   (following pattern from blacksmith)