Converted JobStatuses.usualRoutine to be backed by a decision tree
Instead of condition logic that you need to walk through completely, the
new logic is based on a tree-like object which can be "populated" as
soon as possible and therefore more easily debugged at runtime because
you can inspect the entire state of the world and once and understand
how it decided the next status.

This is a fairly bulky item to make it easier to debug against. It may
be replaced with a lighter weight one in the future for performance
reasons.

TODO (Fetcher):
- Fetcher doesn't seem to start
  - RNIOT is empty at state 0?
- Graphics
- Use embedded request in job behaviour