Refactored status computation and added fetcher (#192)
**Refactor**

This makes it easier to debug the status calculation because you can look at the entire logic chain in a java debugger rather than having to step through a bunch of conditionals.

It might not have the best performance, but right now my goal is reliability. And the performance seems "acceptable".

**Fetcher**

The fetcher takes a stock request (and item which specifies a desired item) from a chest, then they search the rest of town for the desired item and put it in the chest from where they took the stock request. Once there are only two slots left in the target chest, they drop the last item and the stock request.

This was not (fully) implemented using the gold-standard "special-rule" style because I simply couldn't get it to work well. Instead, the game's core codebase has knowledge of the fetcher. I did my best to keep all the fetcher-related code contained within the FetcherHacks class.